Task recorder state machine
Implements a mini-FSM for the recorder task. The recorder has three
states:
  - REC_OFF
  - REC_FILL_QUEUE
  - REC_WRITE_TO_FLASH

Whenever a state change is requested the set_recorder_state triggers
a command and sends it to the recorder task.

In this way, the recorder task is the only one who is actually
accessing the files (except the rec_print_flight CLI command),
thereby removing the need for a mutex over LFS access.