Streams on Top of FIFOs
This patch introduces the concepts of streams and stream groups.

A stream is an abstraction on top of a FIFO that provides a more
reliable way to insert data into the FIFO. Currently FIFOs give up on
writing into their underlying buffer if their mutex is taken by someone
else.

On the other hand, a stream tries to write into the FIFO until it is
successful, or a predefined timeout expires.

--

A stream group is just a wrapper for a pair of input and output streams.

--

Additionally, a "USB stream group" is defined, with a defined timeout of
10ms.

--

All USB communication (CLI and log_* functions) now uses streams. This
enables the removal of the print_mutex and osDelays in log_* functions.
This makes flight log dumping much, much faster.