Add support for printing of generic config.
Currently printing of the config is directly tied to the
global_cats_config variable.

This patch replaces storing the direct pointers to global_cats_config
members by keeping track of their offset, which makes the functions
parameterizable with an arbitrary instance of cats_config_u.

This work is a prerequisite for adding the flight config to the stats
file since we have to load the saved config to a new variable when we
want to print it out.